------------------------------------------------------------------------ r3187 | loera01 | 2010-06-25 13:30:14 -0400 (Fri, 25 Jun 2010) | 34 lines
This change is a contuation of the fix for bug 122174 introduced in change
498416.  The scope of the change has expanded to SQLGetDescRec(),
SQLColumns(), and SQLDescribeCol().

The bulk of this changes follows up the work begun in change 498416, which
defaults SQL_DESC_LENGTH to the column size rather than the octet length.
In most cases, the column size returns the same value as precision, but
diverge for date/time types that include second fractions:  SQL_DESC_LENGTH is
the number of significant digits in the date, and SQL_DESC_PRECISION is the
number of digits past the decimal point.  For decimal types,
SQL_DESC_PRECISION is the number of significant digits, and SQL_DESC_SCALE
is the number of digits past the decimal point,s one might expect.

Other adjustments were required for SQL_DESC_OCTET_LENGTH.  The ODBC spec
requires SQL_DESC_OCTET_LENGTH to represent the number of bytes required to
convert the column to a character string, as opposed to the number of bytes
in the column as retured from the DBMS.

The Length field in tDESCFLD is unused in ODBC except for metadata queries.
This means that SetDescDefaultsFromType() can set Length directly without
causing regression elsewhere.  Ideally, SetDescDefaultsFromType() would set
all IRD lengths correctly right off the bat.  This is not the case
unfortunately.  For cases in which Length or OctetLength are out of spec,
the appropriate values are set in the metatdata functions themselves.

Change 498416 added a function called GetLengthFromType() to set the Length
Field.  I removed this function and merged its capabilities into
SetDescDefaultsFromType().

I incremented the ODBC version number to 3.50.1000.0120.  This change isn't
associated with a patch, but is significant enough to be associated with a
new version number.

------------------------------------------------------------------------
r3188 | stial01 | 2010-06-25 14:15:10 -0400 (Fri, 25 Jun 2010) | 8 lines

B123972, Servicedesk 143851
Alter table fails on a table where the row size was already greater than
the page size.

Description of changes:
Use macro to check for all page type(s) that allow rows spanning pages.

------------------------------------------------------------------------