------------------------------------------------------------------------ r3191 | loera01 | 2010-06-25 18:15:11 -0400 (Fri, 25 Jun 2010) | 16 lines
While preparing a submission related to change bug 122174, I noticed some in
accuracies in the treatment of SQL_DESC_LENGTH, SQL_DESC_PRECISION,
SQL_DESC_SCALE and SQL_DESC_OCTET_LENGTH for interval types.   This change
gives interval types, especially DAY TO SECOND, the same treatement as the
other date/times types.  Of special note is SQL_DESC_PRECISION, which
represents the numbers to the left of the decimal point.

In order to test interval support, it was also necessary to make a change to
ConvertEscAndMarkers().  As it stands, ConvertEscAndMarkers() translates the
escape sequence syntax { interval 'ddd hh:mm:ss[.fff...]' } to
INTERVAL'ddd hh:mm:ss[.fff...]'.   This was perfectly valid syntax when ISO
date/time types was first introduced, but this is no longer the case; the
keyword INTERVAL is no longer supported.  ConvertEscAndMarkers() now omits the
INTERVAL keyword.

------------------------------------------------------------------------