------------------------------------------------------------------------ r3191 | loera01 | 2010-06-25 18:15:11 -0400 (Fri, 25 Jun 2010) | 16 lines
While preparing a submission related to change bug 122174, I noticed some in
accuracies in the treatment of SQL_DESC_LENGTH, SQL_DESC_PRECISION,
SQL_DESC_SCALE and SQL_DESC_OCTET_LENGTH for interval types.   This change
gives interval types, especially DAY TO SECOND, the same treatement as the
other date/times types.  Of special note is SQL_DESC_PRECISION, which
represents the numbers to the left of the decimal point.

In order to test interval support, it was also necessary to make a change to
ConvertEscAndMarkers().  As it stands, ConvertEscAndMarkers() translates the
escape sequence syntax { interval 'ddd hh:mm:ss[.fff...]' } to
INTERVAL'ddd hh:mm:ss[.fff...]'.   This was perfectly valid syntax when ISO
date/time types was first introduced, but this is no longer the case; the
keyword INTERVAL is no longer supported.  ConvertEscAndMarkers() now omits the
INTERVAL keyword.

------------------------------------------------------------------------
r3192 | lunbr01 | 2010-06-25 18:30:12 -0400 (Fri, 25 Jun 2010) | 34 lines

Socket option TCP_NODELAY was added to the Windows tcp_ip network
protocol driver in Ingres 9.3.1, but was defaulted to OFF, primarily
due to concern over recommendation in Microsoft documentation to
leave if OFF in most environments.  Since then, testing has shown a
couple of cases where performance has improved noticeably (~5%) with
the option ON.  All other test cases have shown no difference in
response time...and never a negative impact.  It has been decided to
default the option to ON to get the best "out of the box" performance
on Windows when using Ingres/Net and/or using 'tcp_ip' as the local
IPC.  Two cases where performance was improved with the option ON
were a batch load (series of inserts) and an interactive application
when run using "direct connect" across the network between Windows
machines with II_GC_PROT=tcp_ip.

The original variables in Ingres 9.3.1 for enabling TCP_NODELAY were
II_WINSOCK2_NODELAY and winsock2_nodelay.  This change adds 2 new
variables which are more clear, easier to remember and more
consistent with other config variables related to the 'tcp_ip'
protocol driver.  The new names are Ingres environment variable
II_TCPIP_NODELAY and config.dat parameter tcp_ip.nodelay.  Although
the old names were never officially documented, they were provided
to at least one customer.  For that reason, the old names were kept
also.  If both set, the new names take precedence.

At this time, the config option was not added to the config
utilities (cbf and vcbf).  It is expected (hoped) that the option
will, in fact, never need to be turned off.  If that changes, it
should be added to the config tools.  That would, however, require
reorganizing the protocols config screen a bit as it is currently
generic (across protocols) and only provides the ability to change
status and listen port.  FYI, Unix/Linux always turn on TCP_NODELAY
(no option to turn it off).

------------------------------------------------------------------------