------------------------------------------------------------------------ r3192 | lunbr01 | 2010-06-25 18:30:12 -0400 (Fri, 25 Jun 2010) | 34 lines
Socket option TCP_NODELAY was added to the Windows tcp_ip network
protocol driver in Ingres 9.3.1, but was defaulted to OFF, primarily
due to concern over recommendation in Microsoft documentation to
leave if OFF in most environments.  Since then, testing has shown a
couple of cases where performance has improved noticeably (~5%) with
the option ON.  All other test cases have shown no difference in
response time...and never a negative impact.  It has been decided to
default the option to ON to get the best "out of the box" performance
on Windows when using Ingres/Net and/or using 'tcp_ip' as the local
IPC.  Two cases where performance was improved with the option ON
were a batch load (series of inserts) and an interactive application
when run using "direct connect" across the network between Windows
machines with II_GC_PROT=tcp_ip.

The original variables in Ingres 9.3.1 for enabling TCP_NODELAY were
II_WINSOCK2_NODELAY and winsock2_nodelay.  This change adds 2 new
variables which are more clear, easier to remember and more
consistent with other config variables related to the 'tcp_ip'
protocol driver.  The new names are Ingres environment variable
II_TCPIP_NODELAY and config.dat parameter tcp_ip.nodelay.  Although
the old names were never officially documented, they were provided
to at least one customer.  For that reason, the old names were kept
also.  If both set, the new names take precedence.

At this time, the config option was not added to the config
utilities (cbf and vcbf).  It is expected (hoped) that the option
will, in fact, never need to be turned off.  If that changes, it
should be added to the config tools.  That would, however, require
reorganizing the protocols config screen a bit as it is currently
generic (across protocols) and only provides the ability to change
status and listen port.  FYI, Unix/Linux always turn on TCP_NODELAY
(no option to turn it off).

------------------------------------------------------------------------
r3193 | coomi01 | 2010-06-28 05:00:11 -0400 (Mon, 28 Jun 2010) | 5 lines

Detect case where no table copying is to be performed,
and where journaling is required. Then output (with)
journaling in the create tables statement.

------------------------------------------------------------------------