------------------------------------------------------------------------ r3193 | coomi01 | 2010-06-28 05:00:11 -0400 (Mon, 28 Jun 2010) | 5 lines
Detect case where no table copying is to be performed,
and where journaling is required. Then output (with)
journaling in the create tables statement.

------------------------------------------------------------------------
r3194 | schka01 | 2010-06-28 07:15:18 -0400 (Mon, 28 Jun 2010) | 11 lines

The identity column code shares a few routines and
parser rules with CREATE SEQUENCE.  The pss_object
for CREATE SEQUENCE is a PSY_CB, but for create table,
it's a QEU_CB.  Make sure that there are no PSY_CB
references from any routine or parser rule that can be
called by create table with identity columns.  Get rid
of the psy_seqflags flag word, as there was only one
really necessary flag in it, and use a temporary flag
in the iisequence tuple to replace it.

------------------------------------------------------------------------