------------------------------------------------------------------------ r3196 | gupsh01 | 2010-06-28 09:00:09 -0400 (Mon, 28 Jun 2010) | 18 lines
On Windows 32 bit creating a user with accessdb fails
with E_US10CE error. This was observed on installing
II 10.0.0 (int.w32/125) on Windows XP and using the
accessdb program to create a new user.

The problem occurs because Windows optimizer tries to
save the parameters passed to the user_create() routine
eg x_profile, x_exp_date, x_group in cpu registers,
as against saving them on the stack. However in
subsequent call to ERget() the pointer value in the
register gets overwritten, resulting in corruption of
data used to form the query strings for updating the
new user.

Turned off optimization for this file user.qsc to fix
the problem.

------------------------------------------------------------------------