------------------------------------------------------------------------ r3199 | drivi01 | 2010-06-28 23:30:10 -0400 (Mon, 28 Jun 2010) | 11 lines
QA is unable to run any of the handoffqa tests on Windows 2008 machine.
This is due to the fact that sepchild fails to lower elevation on the
testing commands.  GetTokenInformation for linkedtoken with stripped
privileges returns error 1312.

This change fixes the problem by retrieving current token type
before attempting to retrieve linked token.  Linked token
only exists if the elevation type is full, on Windows 2008
the elevation type is default and doesn't contain a linked token.

------------------------------------------------------------------------
r3200 | drivi01 | 2010-06-29 01:20:52 -0400 (Tue, 29 Jun 2010) | 3 lines

Resolve build issues in VSDA as a result of change 505605.

------------------------------------------------------------------------