------------------------------------------------------------------------ r3206 | coomi01 | 2010-06-29 10:15:28 -0400 (Tue, 29 Jun 2010) | 4 lines
The test for choosing to flag journaling should be the opposite to
that used for copying.

------------------------------------------------------------------------
r3207 | kiria01 | 2010-06-29 12:04:44 -0400 (Tue, 29 Jun 2010) | 6 lines

The parse tree has been altered for comparison operators that involve either
'column OP const' or 'const OP column' where const is or might be a unicode
constant often derived from '?' parameters. In these cases a call is made to
psl_mk_const_similar() to ensure that the unicode constant is down graded to
match the type of the column.

------------------------------------------------------------------------