------------------------------------------------------------------------ r3207 | kiria01 | 2010-06-29 12:04:44 -0400 (Tue, 29 Jun 2010) | 6 lines
The parse tree has been altered for comparison operators that involve either
'column OP const' or 'const OP column' where const is or might be a unicode
constant often derived from '?' parameters. In these cases a call is made to
psl_mk_const_similar() to ensure that the unicode constant is down graded to
match the type of the column.

------------------------------------------------------------------------
r3208 | loera01 | 2010-06-29 13:15:11 -0400 (Tue, 29 Jun 2010) | 20 lines

The minimum and maximum scales for TIMESTAMP were listed as zero.  The maximum
scale shouuld have been 9.

Research on TIME and INTERVAL DAY TO SECOND data types showed -1 for minimum
and maximum scales.  These should both return a minimum scale of zero and a
maximum scale of 9.  Research on the fix for bug 122174 indicated that the
column size was incorrect for TIMESTAMP, TIME, and interval types.

Modified the column size in tdataTypeInfo according to the ODBC specs for
date/time data types:

Data Type                     Column Size

DATE                          10 (characters in YYYY-MM-DD)
TIMESTAMP                     29 (20 + max precision of 9)
TIME                          18 (9 + max precision of 9)
INTERVAL YEAR TO MONTH         3 (3 + max precision of 0)
INTERVAL DAY TO SECOND        19 (10 + max precision of 9)

------------------------------------------------------------------------