------------------------------------------------------------------------ r3209 | kiria01 | 2010-06-29 13:24:48 -0400 (Tue, 29 Jun 2010) | 7 lines
Correct the maintenance of qual_depth with PST_ROOT otherwise
subsequent PST_ROOT nodes from unions appear at ever deeper parse
depths resulting in setting bits out of range in in_WHERE and
has_COUNT and on down growing stacks this will corrupt stack as
seen in stack trace.

------------------------------------------------------------------------
r3210 | hanje04 | 2010-06-29 14:00:11 -0400 (Tue, 29 Jun 2010) | 9 lines

Previous fix for bug 123941, resolved the issue for RHEL 4 and worked for
RHEL 5 but caused issues on RHEL 6. The root of the issue is difference in
the implementation of 'action' between RHEL 4,5 & 6 which means it's
impossible define the INGSTART/INGSTOP variables so that the commands run on
all 3. To resolve this the commands are now written out to a temporary file
which is run as a shell script by "action". This removes any need for variable
deferencing in the command its self and seems to work everywhere.

------------------------------------------------------------------------