------------------------------------------------------------------------ r3210 | hanje04 | 2010-06-29 14:00:11 -0400 (Tue, 29 Jun 2010) | 9 lines
Previous fix for bug 123941, resolved the issue for RHEL 4 and worked for
RHEL 5 but caused issues on RHEL 6. The root of the issue is difference in
the implementation of 'action' between RHEL 4,5 & 6 which means it's
impossible define the INGSTART/INGSTOP variables so that the commands run on
all 3. To resolve this the commands are now written out to a temporary file
which is run as a shell script by "action". This removes any need for variable
deferencing in the command its self and seems to work everywhere.

------------------------------------------------------------------------
r3211 | schka01 | 2010-06-29 19:30:12 -0400 (Tue, 29 Jun 2010) | 8 lines

Table procedure validation occurs in the main session thread.
If the query plan is a parallel plan, we mustn't use the DSH
created for tproc validation for an actual tproc node, if the
node is under an exchange.  For that case, go ahead and create
another DSH for the tproc node, which will occur in the child
thread context.

------------------------------------------------------------------------