------------------------------------------------------------------------ r3211 | schka01 | 2010-06-29 19:30:12 -0400 (Tue, 29 Jun 2010) | 8 lines
Table procedure validation occurs in the main session thread.
If the query plan is a parallel plan, we mustn't use the DSH
created for tproc validation for an actual tproc node, if the
node is under an exchange.  For that case, go ahead and create
another DSH for the tproc node, which will occur in the child
thread context.

------------------------------------------------------------------------
r3212 | bonro01 | 2010-06-29 20:31:45 -0400 (Tue, 29 Jun 2010) | 3 lines

Update Documentation for Ingres 10.0

------------------------------------------------------------------------