------------------------------------------------------------------------ r3224 | whiro01 | 2010-07-01 12:30:10 -0400 (Thu, 01 Jul 2010) | 26 lines
Change 505643 introduced code in the "ingres!tooldev" area to create a
C/C++ Tools API as the basis for the new Ingres Administration Tools.

This Tools API needs access to certain low-level Ingres data
structures.  But, since the API code is mainly written in C++, it is
more sensitive to the presence / absence of function prototypes than
regular C code.  Thus, the inclusion of "nmlocal.h" in order to access
these data structures causes a compilation error because the prototype
for "NMstIngAt" is duplicated (almost) in two different headers and
the difference between the two prototypes causes the compiler to
complain.

The header file "nmcl.h" used for both Windows and UNIX contains a
full prototype for "NMstIngAt()" (at least when CL_PROTOTYPED is
defined, which it always is these days).  The "nmlocal.h" file,
however, for Windows and UNIX contains only a partial (return value
only) prototype for this same function.  In C this is not a problem,
but in C++ (which is used for the new Tools API) this creates an error
because the language requires full prototypes be defined for all
functions and the two prototypes do not agree.  So, the partial one in
"nmlocal.h" is removed in favor of the complete one in "nmcl.h" (which
is always included in the files that include "nmlocal.h", except for
the UNIX version of "nmgtat.c" which doesn't reference "NMstIngAt()"
anyway).

------------------------------------------------------------------------
r3225 | hanje04 | 2010-07-01 18:45:10 -0400 (Thu, 01 Jul 2010) | 4 lines

Use the installation ID of the instance being upgraded/modified when generating
the service script name. NOT the next default instID.

------------------------------------------------------------------------