------------------------------------------------------------------------ r3243 | schka01 | 2010-07-05 09:00:16 -0400 (Mon, 05 Jul 2010) | 11 lines
Make sure that the proper partition DMR_CB is set up when
doing scrollable cursor fetches.  Make sure that when a
row is fetched from the QEP for a keyset cursor, the TID
is copied to the temp-file result area which is where
ahd_tidrow and ahd_tidoffset are pointing.  Cleanup up
RUP and RDEL generation a little so that they aren't
dependent on runtime activities to set up the right
DMR_CB;  the proper CB can be figured out at cursor
compile (open) time.

------------------------------------------------------------------------