------------------------------------------------------------------------ r3245 | drivi01 | 2010-07-06 01:21:06 -0400 (Tue, 06 Jul 2010) | 9 lines
Upgradedb on star user databases is failing on Windows when
upgrading from 9.2.1 release to 10.0.0.

This change switches ServerPostInstallation with
StarPostInstallation, so that star server is setup
before the server, so that when it comes to run
upgradedb on user database star server is up and running.

------------------------------------------------------------------------
r3246 | smeke01 | 2010-07-06 04:30:39 -0400 (Tue, 06 Jul 2010) | 36 lines

Review Board IP  http://reviewboard.ingres.prv/r/4698

Review Board IP  http://reviewboard.ingres.prv/r/4698

Cross integration of change 504829

To make sure that the same joinid(s) is/are used for any repeat copies of the
AGHEAD WHERE clause we save the starting value of header->pst_numjoins. We
pass the value in a separate write parameter so that it can be incremented as
required by later functions.  These functions also handle incrementing
header->pst_numjoins. For subsequent copies of the AGHEAD WHERE clause, we
reset the write parameter to the saved value of header->pst_numjoins.

Cross integration of changes 505309 and 505713

Ensure that certain global maps in opa_notex_transform() and
opa_notin_transform() are updated only for the first copy of the AGHEAD tree.
'global maps' in this context means:
- The global rangetable pst_outer_rel and pst_inner_rel maps
- The table map for the root of the query ('outer_root')

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------