------------------------------------------------------------------------ r3250 | smeke01 | 2010-07-07 04:30:27 -0400 (Wed, 07 Jul 2010) | 33 lines
Review Board IP  http://reviewboard.ingres.prv/r/4712

Cross integration of change 505222

Check opv_mbf.opb_count and opv_mbf.opb_bfcount for the relation passed
in to opn_prtree() to see whether the relation:
(1) has a multi-attribute key with matching attributes in the query
(opb_count > 1).
(2) has a multi-attribute key with at least one matching boolean factor in the
query (opb_bfcount > 0).
If either is false then there is either no multi-attribute key, or a file scan
is preferable to using the multi-attribute key and we therefore prevent
oph_origtuples / oph_odefined from being set for the histograms of the
multi-attribute key. This in turn prevents the 'delayed multi-attribute' code
in opn_jtuples() from being skewed by an unreasonably high number for
oph_origtuples (up to 50% of the number of rows in the base table).

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------
r3251 | kiria01 | 2010-07-07 06:45:15 -0400 (Wed, 07 Jul 2010) | 11 lines

String-numeric compare failing in IN-list context

The string numeric context is usually detected in the resolver
but optimising the IN lists was obscuring the possibility having
forced the datatype of the constants to match the column type.
The specific case of string column expression with a numeric
IN-list is now checked for and if found, numericnorm functions
are injected to retain the string-numeric behavior.

------------------------------------------------------------------------