------------------------------------------------------------------------ r3251 | kiria01 | 2010-07-07 06:45:15 -0400 (Wed, 07 Jul 2010) | 11 lines
String-numeric compare failing in IN-list context

The string numeric context is usually detected in the resolver
but optimising the IN lists was obscuring the possibility having
forced the datatype of the constants to match the column type.
The specific case of string column expression with a numeric
IN-list is now checked for and if found, numericnorm functions
are injected to retain the string-numeric behavior.

------------------------------------------------------------------------