------------------------------------------------------------------------ r3253 | hanje04 | 2010-07-07 14:15:12 -0400 (Wed, 07 Jul 2010) | 14 lines
Changes to the GUI installer for Ingres 10 introduced another command into the
command line generated at install time to run the setup, but didn't allocate
any more space in the buffer to store it. When adding a package to an existing
instance we overrun the allocated buffer sufficiently to cause an exception:

    *** glic detected *** ./bin/inginstgui: free(): invalid next size (normal): 0xf5d01200 ***

To avoid this, include the extra command when calculating the space need for
the command line. Also add new gipAppendCmdLine() function and use that instead
of STcat() so we can keep track of available space as we go. If we are going to
exceed the allocated space, alert the user and exit cleanly.

------------------------------------------------------------------------