------------------------------------------------------------------------ r3256 | stephenb | 2010-07-07 22:03:35 -0400 (Wed, 07 Jul 2010) | 35 lines
Release Notes:
rns.start
     (GENERIC)
     124024
     When running insert statements in a batch that use the insert to copy optimization,
     on heap tables, if very small batches are used, the data will be inserted more
     slowly than a non-batched insert; this defeats the point of using batches.
rns.end

Description of changes:
desc.start
The batched insert to copy optimization will utilize the standard QEF "copy" code to load the table,
this code will try to use the DMF loader interface (dmr_load) if it can. In most cases, when using
the optimization, this will not be possible because the table will already contain data. Heap tables
are the exception to this since dmr_load can be used on these tables when there are existing records.
dmr_load is optimised for load large amounts of data, and when it is used for very small numbers of rows
(less than some number between 10 and 100 depending on row size and machine speed), it is less efficient
than simply calling dmr_put.

This change creates a row buffer for the insert to copy optimization and adds the ability to collect
a buffer full of rows before calling the QEF copy code to load the data. If there is a very small number
of rows, we know that before we call copy, and we will set a flag to indicate this to QEF.
We can now make some smart decisions about how to execute the copy. The QEF copy code has been
changed so that it will not use dmr_load, or create threads to execute a copy when it knows that the size
of the copy is very small. In addition, the QEF copy code must now deal with the fact that each row in the
multi-row copy buffer may have a different external size and format. Code has been added to track this
information in PSF when reading the input query, and QEF contains special processing for
copy requests that employ the optimization which will cope with the fact that the row size
and format might be different for each row.

Related Service Desk Issues: NONE

Build Considerations: NONE
desc.end
------------------------------------------------------------------------
r3257 | hanal04 | 2010-07-08 06:00:10 -0400 (Thu, 08 Jul 2010) | 19 lines

Cross integration of change 505871

op150 trace output incorrectly dumps the first parameter of an
ADE_CALCLEN instruction as a DBV Datavalue. In fact the DBV has been
overloaded with an ADI_LENSPEC. This change adds ade_prlenspec()
to dump an ADI_LENSPEC that has been stored in a DBV. We now call
ade_prlenspec() to dump the first operand of an ADE_CALCLEN
instruction.

Exmple output after the change:

70: ADE_CALCLEN(11[0]{ADI_O1} , 11[0]{DB_VCH_TYPE(-1)} ,
10[0]{DB_VCH_TYPE(-1)} );

d0: ADE_CALCLEN(12[0]{ADI_VARCTOU} , 12[0]{DB_NCHR_TYPE(-1, 0)} ,
11[0]{DB_VCH_TYPE(-1)} );

------------------------------------------------------------------------