------------------------------------------------------------------------ r3261 | abbjo03 | 2010-07-08 13:45:13 -0400 (Thu, 08 Jul 2010) | 34 lines
Cross-integrate change 505885 from ingres2006r3.
The change of 17 June 2009 to excatch.c introduced the use of longjmp
into EXcatch, which is used as a VMS condition handler.  The HP
documentation states:

"The longjmp function will not work if invoked from nested signal
handlers. The result of the longjmp function, when invoked from a signal
handler that has been entered as a result of an exception generated in
another signal handler, is undefined."

Most of the uses of EXsignal(EX_JNP_LJMP) in Ingres code result in a
single level unwind.  However, the 'set joinop timeoutabort' causes a
two level unwind and thus longjmp cannot be used in that situation.

    Revert most of the June 2009 excatch.c change.  However, instead of
    using the ICB "valid" masks for restoring registers prior to the
    unwind, use a more conservative approach restoring primarily the
    "preserved" registers, similar to what we do on Alpha VMS.

    Change the EX_CONTEXT structure to include both the ICB (jmp_buf)
    and a pointer to it.  Although undocumented, HP informed us that the
    start of the ICB is always adjusted, even when already
    octaword-aligned.

    Change the EXdeclare macro to use lib$i64_get_curr_invo_context
    instead of the undocumented decc$setjmp1.  Change EXsetup to store
    and adjust the pointer to the ICB, and initialize the block.

    Compile back/opf/opn/opnenum.c and front/tm/monitor/tmmain.c with
    NOOPTIM on IA64 VMS as optimization prevents proper register
    restoration.

------------------------------------------------------------------------
r3262 | drivi01 | 2010-07-09 07:15:32 -0400 (Fri, 09 Jul 2010) | 10 lines

User is seeing unresolved symbol security_cookie during linking of test
applications.  This is due to the fact that MS compiler turns on /GS
flag by default now and builds security_cookie checks in the beginning
and the end of the functions.

This change adds libcmt.lib to the link lines and also adds sepcob.bat
to the handoffqa tools for QA.

------------------------------------------------------------------------