------------------------------------------------------------------------ r3265 | wanfr01 | 2010-07-09 15:15:10 -0400 (Fri, 09 Jul 2010) | 18 lines
Issue 141863
Bug 123949 - E_RD0125 consistency check happening along with
occasional E_CL250A errors.

For this particular client, they are getting E_DM9043 timeouts due to
modify statements running against tables with heavy concurrent usage.
If the opf_active_timeout dbms parameter is not set to a fairly high number
(connect limit+2), you can get E_DM9043 control lock timeouts in dmt_show.

When opv_parser fetches table information via rdf_gdesc, if it gets an error,
then it should discard the fetched information.  The information can be
incomplete and have uninitialized values.

At the moment it currently uses what it has, and you get semaphore errors
and/or segvs.  The code currently assumes that if the global range variable
is not null then it is 'complete'.

------------------------------------------------------------------------
r3266 | loera01 | 2010-07-09 16:00:11 -0400 (Fri, 09 Jul 2010) | 21 lines

The BS_DRIVER global table defines the functions for each protocol,
including a portmapper function that converts the symbolic port plus the
subport (instance number) to a port value.  The fix for bug 120552 adds an
extra output argument string that describes the listen symbol according to
the format.  Thus, a symbolic port of "RT" returns  25760 and RT for the
first instance, 25761 and RT1 for the second, etc. On VMS, DECNet and
TCP/IP are supported, so the portmapper function required changes for both
the DECNet and TCP/IP versions, which are GC_decnet_addr() and
GC_tcp_addr(), respectively.  SNA_LU0, TCP_WOL and ASYNC have portmapper
functions as well, but are not addressed in this change because these
protocols are now obsolete and should be removed.

For GC_tcp_addr(), I copied the code verbatim from the Unix version,
BS_tcp_addr().  On DECNet, numeric ports are not used, so the fix for
120552 doesn't apply.  In the DECNet case, I simply copied the base port to
the returned symbolic port.  Thus, the first instance of the GCC returns an
address of II_GCCRT as the address and II_GCCRT_0 for the port, the second
instance as II_GCCRT and II_GCCRT_1, etc.

------------------------------------------------------------------------