------------------------------------------------------------------------ r3268 | coomi01 | 2010-07-12 04:30:13 -0400 (Mon, 12 Jul 2010) | 5 lines
Check for any attempt to adjust a location for iidbdb
and prevent it from happening by issuing a (new) error
message E_US18E8.

------------------------------------------------------------------------
r3269 | coomi01 | 2010-07-12 04:45:31 -0400 (Mon, 12 Jul 2010) | 7 lines

When examining histogram key values for duplicates, do not compare
the TM string values rather compare the histogram keys directly.
If the keys are binary, they may contain c-null bytes in which
case the string comparison will not look in the tail of the key
for differences giving rise to a suprious error message.

------------------------------------------------------------------------