------------------------------------------------------------------------ r3273 | schka01 | 2010-07-12 09:15:24 -0400 (Mon, 12 Jul 2010) | 39 lines
Define a new standard compression type that compresses
BYTE and VARBYTE columns.  The algorithm is analogous to
CHAR and VARCHAR:  BYTE has trailing zeros trimmed off,
and the remainder stored as if it were VARBYTE.  VARBYTE
is stored as the exact value length as indicated by the
length count.

The existing TCB_C_STANDARD compression type was renamed
to TCB_C_STD_OLD, as it must remain in order to be able to
handle existing tables created before this change.

From a user perspective, there is no way to tell between the
two other than by looking at iirelation.relcomptype. (Old
standard is 1 and new standard is 3.)  iitables.is_compressed
returns Y for both old and new compression.  Front-end
utilities such as copydb or unloaddb will report normal
data compression for either new or old style.

Any table modified to compression=(data) or equivalent gets the
new style compression, regardless of what compression it used
to have.  This applies to modify to reconstruct as well.
There are two exceptions to this rule.

1. BTREE key compression always uses old compression.  There
has historically been confusion about how index compression is
specified inside of DMF, so it's best to leave it alone for now.

2. Catalog and in particular core catalog compression remains
old style compression.  While it would be nice to be able to
compress the relenckey in iirelation, dm2d has no current ability
to deal with multiple core-catalog compression types and now is
not a good time to change it.  The alternative of assuming that
core catalogs always use new compression would invalidate all
existing 10.0 installations, requiring them to do a full
unload / reinstall / reload.  That doesn't seem like a good
idea either.  The core catalog type can be changed for some
future release, when we're going to do upgradedb anyway.

------------------------------------------------------------------------
r3274 | drivi01 | 2010-07-12 12:30:15 -0400 (Mon, 12 Jul 2010) | 8 lines

.Net Data Provider and Documentation stand-alone installers do not display
the license dialog or have capability to fetch the license at runtime.

This change brings .NET Data provider and Documentation installers inline
with Ingres installer where the licenses can be fetched at runtime
and license dialog is displayed when stand-alone wizard is used.

------------------------------------------------------------------------