------------------------------------------------------------------------ r3275 | schka01 | 2010-07-12 14:45:10 -0400 (Mon, 12 Jul 2010) | 26 lines
Remove the "direct-IO mandatory" code that used to
turn on direct IO at the lowest level (IIdio_open)
without the caller's knowledge.
Earlier 10.0 DI changes removed DI code that
automatically reallocated and aligned the read/write
area if the fd was opened O_DIRECT and the caller's
buffer was unaligned.  This auto realignment was
very expensive;  the new protocol is that callers
should be aware of direct IO, and should request
it only when they are prepared to supply a properly
aligned and sized buffer.  The direct-IO-mandatory
business bypassed all that, and so installs failed
with unaligned buffer errors.

Since the direct-IO-mandatory code was an attempt
at optimization, rather than a correctness
requirement, it has been removed.  Users installing
on filesystems that need direct IO for performance
can use the direct_io, direct_io_log, and
direct_io_load configuration settings.

Any filesystem type that requires direct IO for
correct Ingres operation will be banned outright.
No instances of this are known at present.

------------------------------------------------------------------------