------------------------------------------------------------------------ r3277 | stephenb | 2010-07-12 19:56:54 -0400 (Mon, 12 Jul 2010) | 31 lines
Release Notes:
rns.start
     124058
     (GENERIC)
     Executing statements in a batch through JDBC that do not utilize the copy optimization
     may sometimes case a hang if the batch size is too large.
rns.end

Description of changes:
desc.start
The problem is caused by an architectural issue. When processing batches, the DAS server will continue
to send requests without reading the receive queue until the batch is done, whilst the DBMS will call
GCA_SEND with GCA_NOT_EOG after processing each request. Even though GCA_NOT_EOG is set, GCA will
eventually send the responses to the DAS server when the GCA packet fills up. Because the DAS server
is not reading the responses, and because GCA_SEND is not fully asynchronous, the DBMS will eventually
hang in GCA_SEND waiting for the DAS server to read the responses. Meanwhile the DAS server will continue
to send requests and will eventually hang in GCA_SEND waiting for the DBMS to read the requests.
We have a send/receive deadlock.

The reason this problem is less likely to occur when the copy optimization is used is because that
optimization causes the DBMS to save up the responses until the copy is done (in case the copy fails), and
so the DBMS is not sending the responses back after processing each request. This change extends the
behaviour to all batch processing. Now, the DBMS will queue the responses for the entire batch, and will not
call GCA_SEND until the batch is done; it is able to do this because the user should not expect any response
from a batch until the program indicates that the batch has completed.

Related Service Desk Issues: NONE

Build Considerations: NONE
desc.end
------------------------------------------------------------------------